tests/requirements.txt
yandex-taxi-testsuite[postgresql-binary] >= 0.1.7.2
grpcio
grpcio-tools
